test: increase coverage of WalletDelegate
